requirements.txt
flask
python-dotenv
ccxt